app/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-split-settings==1.1.0
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
sqlparse==0.4.2
